improved Kiwi.tokenize & added Kiwi.splitIntoSents
